Support GHC 8.10
