+++ microlens-ghc/microlens-ghc.cabal
@@ -27,6 +27,7 @@ tested-with:         GHC==7.4.2
                      GHC==8.4.4
                      GHC==8.6.5
                      GHC==8.8.1
+                     GHC==8.10.1
 
 source-repository head
   type:                git
