+++ microlens-th/microlens-th.cabal
@@ -25,6 +25,7 @@ tested-with:         GHC==7.4.2
                      GHC==8.4.4
                      GHC==8.6.5
                      GHC==8.8.1
+                     GHC==8.10.1
 
 source-repository head
   type:                git
@@ -40,7 +41,7 @@ library
                      , containers >=0.4.0 && <0.7
                      , transformers
                      -- lens has >=2.4, but GHC 7.4 shipped with 2.7
-                     , template-haskell >=2.7 && <2.16
+                     , template-haskell >=2.7 && <2.17
                      , th-abstraction >=0.2.1 && <0.4
 
   ghc-options:
